remove old tap page, added skeleton for new collect page
